move main into LessEngine
